+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Changed the structure of the `validate` property setting to be an array with named options.
 - The unique constraint is now a validation rule specified as: `['validate' => ['unique', 'column' => 'email']]`.
 - Deprecated the `password_php` validation rule in favor of `password`.
+- Callable validation rules must now be specified as: `['validate' => ['callable', 'fn' => ...]]`.
 
 ### Fixed
 - Use a strict equality check when casting an empty string to null on a nullable property. Previously this would check for a falsey value.